% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary.functions.R
\name{acomb}
\alias{acomb}
\title{Combines 2d arrays into 3-d arraysMeal size of food in g}
\usage{
acomb(...)
}
\description{
This combines the results from each individual simulation "outer loop" into 3-d array.
}
