% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary.functions.R
\name{khf}
\alias{khf}
\title{Annualizes the values of risk to one number for entire year, when
individual daily values are available}
\usage{
khf(x)
}
\description{
Annualizes the values of risk to one number for entire year, when
individual daily values are available
}
