% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.calcs.R
\name{parallel.calcs}
\alias{parallel.calcs}
\title{Final calculations, run in parallel using "foreach".}
\usage{
parallel.calcs(number.simulations, population.size, number.days,
  proportion.secretor.positive, opts1, opts2, output.name, ...)
}
\arguments{
\item{number.simulations}{number of total simulations to run}

\item{population.size}{size of population in each simulation}

\item{number.days}{number of days to run for each individual}

\item{proportion.secretor.positive}{proportion of people NoV Sec Pos}

\item{opts1}{options for outer loop to speed up run speed}

\item{opts2}{options for inner loop to speed up run speed}
}
\description{
Final calculations, run in parallel using "foreach".
}
