% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary.functions.R
\name{prob.infection.beta.poisson}
\alias{prob.infection.beta.poisson}
\title{Probability of getting an infection from pathogen dose, beta-poisson D:R model}
\usage{
prob.infection.beta.poisson(probability.infection.beta,
  probability.infection.alpha, ...)
}
\arguments{
\item{probability.infection.beta}{beta parameter for beta.poisson model}

\item{probability.infection.alpha}{alpha parameter for beta.poisson model}
}
\description{
The probability of getting an infection from the pathogen, given the daily dose, using
the beta.poisson model.
}
