% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.dose.R
\name{total.dose}
\alias{total.dose}
\title{Total Dose}
\usage{
total.dose(run.oyster.model, pathogen.viable.variability.switch, ...)
}
\arguments{
\item{run.oyster.model}{Boolean variable to turn oyster model on or off}

\item{pathogen.viable.variability.switch}{Boolean variable to varible viability
on or off.}
}
\description{
Calculate the total pathogen dose, sum of the four food groups consumed.
}
